rename ArbeidsrettetAktivitet -> RegisterAktivitet
