+++ src/frontend/api/api.ts
@@ -2,11 +2,8 @@ import axios from 'axios';
 import { v4 as uuidv4 } from 'uuid';
 
 import Environment from './Environment';
-import {
-    ArbeidsrettedeAktiviterFraBackend,
-    ArbeidsrettetAktivitet,
-} from '../typer/arbeidsrettetAktivitet';
 import { Person } from '../typer/person';
+import { ArbeidsrettedeAktiviterFraBackend, RegisterAktivitet } from '../typer/registerAktivitet';
 import { Stønadstype } from '../typer/stønadstyper';
 import { Kvittering } from '../typer/søknad';
 
@@ -28,7 +25,7 @@ export const hentPersonData = (): Promise<Person> => {
         .then((response) => response.data);
 };
 
-export const hentArbeidsrettedeAktiviteter = (): Promise<ArbeidsrettetAktivitet[]> => {
+export const hentArbeidsrettedeAktiviteter = (): Promise<RegisterAktivitet[]> => {
     return axios
         .get<ArbeidsrettedeAktiviterFraBackend>(
             `${Environment().apiProxyUrl}/aktivitet`,
